Add support for getting current tag to `GitCLI`
